requirements.txt
streamlit==1.26.0
requests==2.31.0
pillow==10.0.0
numpy==1.25.2
